Adding the Candle.
We've also added a new capability: matrix

If you want to send Tile messages to a device, you should check for
matrix rather than chain.

This is because the Candle supports a matrix of zones but does not have
a chain of devices attached to it.

The LIFX Tile is the only product that currently has a chain of devices.